Replace PulseAudio with SDL2 audio, fix VM arg passing and background music hang
Sound backend: PulseAudio and ALSA both failed to produce audio through
PipeWire's emulation layers. Replaced with SDL2 audio which talks to
PipeWire natively. SDL2 pulls samples from dma.buffer via callback --
no push thread, no SNDDMA_Submit work needed.

Background music infinite loop (snd_dma.c): S_UpdateBackgroundTrack has
a while loop that resamples music into the raw buffer. When dma.speed
(44100) exceeds the music file sample rate (22050), the integer division
`bufferSamples * rate / speed` rounds to zero once bufferSamples reaches
1. Zero fileSamples means zero progress, infinite loop, game hangs.
Never triggered in original Q3 because OSS negotiated 22050 Hz matching
the music files. Fixed with `if (!fileSamples) break`.

VM fixes for x86_64 correctness:
- VMA macro: use VM_ArgPtr(args[x]) instead of raw (void*)args[x] cast;
  QVM args are data segment offsets, not pointers
- VM_ArgPtr/VM_ExplicitArgPtr: take intptr_t, DLL mode passes pointers
  through directly instead of adding dataBase offset
- VM_Call: extract varargs before dispatch -- on x86_64 variadic args
  live in registers, not contiguous on stack after the named param
- VM_CallCompiled stub: delegate to interpreter with proper args array
- OP_BLOCK_COPY: cast src/dest to intptr_t not int for alignment check

+++ code/qcommon/qcommon.h
@@ -324,8 +324,8 @@ int		QDECL VM_Call( vm_t *vm, int callNum, ... );
 
 void	VM_Debug( int level );
 
-void	*VM_ArgPtr( int intValue );
-void	*VM_ExplicitArgPtr( vm_t *vm, int intValue );
+void	*VM_ArgPtr( intptr_t intValue );
+void	*VM_ExplicitArgPtr( vm_t *vm, intptr_t intValue );
 
 /*
 ==============================================================
